Linting

+++ app/uploaders/polymorphic_uploader.rb
@@ -107,8 +107,6 @@ class PolymorphicUploader < CarrierWave::Uploader::Base
 
   def delete_tmp_dir(_new_file)
     # make sure we don't delete other things accidentally by checking the name pattern
-    if @cache_id_was.present? && @cache_id_was =~ /\A[\d-]*\z/
-      FileUtils.rm_rf(File.join(cache_dir, @cache_id_was))
-    end
+    FileUtils.rm_rf(File.join(cache_dir, @cache_id_was)) if @cache_id_was.present? && @cache_id_was =~ /\A[\d-]*\z/
   end
 end
